+++ traktor/commands/timer.py
@@ -8,10 +8,6 @@ from traktor.decorators import error_handler
 from traktor.models import db, Entry, Report
 
 
-app = typer.Typer(name="timer", help="Start/stop timer and reports.")
-
-
-@app.command()
 @error_handler
 def start(project: str, task: Optional[str] = typer.Argument(None)):
     """Start the timer."""
@@ -22,7 +18,6 @@ def start(project: str, task: Optional[str] = typer.Argument(None)):
         )
 
 
-@app.command()
 @error_handler
 def stop():
     """Stop the timer."""
@@ -32,7 +27,6 @@ def stop():
         )
 
 
-@app.command()
 @error_handler
 def status():
     """See the current running timer."""
@@ -42,7 +36,6 @@ def status():
         )
 
 
-@app.command()
 @error_handler
 def today():
     """See today's timers."""
@@ -52,10 +45,12 @@ def today():
         )
 
 
-@app.command()
 @error_handler
-def report(days: int = typer.Argument(default=365, min=1)):
-    """See the current running timer."""
+def report(days: int = typer.Argument(default=0, min=0)):
+    """See the current running timer.
+
+    If days is 0 that means whole history.
+    """
     with db.session() as session:
         output(
             model=Report, objs=engine.report(session=session, days=days),